langchain_rebuild/requirements.txt
# LangChain MindSearch 重构项目依赖

# 核心框架
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# Web 框架
fastapi>=0.104.0
uvicorn>=0.24.0
sse-starlette>=1.6.5
websockets>=12.0

# 数据处理
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 异步处理
aiohttp>=3.9.0
aiofiles>=23.2.0

# 搜索工具
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 数据存储
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# 环境配置
python-dotenv>=1.0.0

# 日志和监控
loguru>=0.7.0
prometheus-client>=0.19.0

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# 中文分词
jieba>=0.42.1

# 可选：向量数据库
# chromadb>=0.4.0
# faiss-cpu>=1.7.0

# 可选：本地模型支持
# transformers>=4.35.0
# torch>=2.1.0